# Implemeting a simple, straightforward bitwise les than protocol