Upon reviewing the BitwiseLessThan protocol (#206), I was struck by how incredibly complex the logic was. I found it very difficult to understand what the paper was saying, and why things needed to be that complicated.

So I thought I would implement a really dumb, straightforward way of doing things and compare performance.

# Comparing two 8-bit numbers
- Old: 44 multiplications
- New: 28 multiplications
36% reduction

# Comparing two 32-bit numbers
- Old: 176 multiplications
- New: 124 multiplications
30% reduction

In general, this "dumb" approach requires 4*(L - 1) multiplications.

In a debug build, when I just run `cargo test` this makes the "manual only" tests run much faster:

# Running the `cmp_all_fp31` test
- Old: 28.65 seconds
- New: 18.64 seconds
35% reduction

# Running the `cmp_random_32_bit_prime_field_elements` test
- Old: 116.54 seconds
- New: 70.74 seconds
39% reduction